+++ Source/Scene/Scene.js
@@ -30,7 +30,8 @@ define([
         './FrameState',
         './OrthographicFrustum',
         './PerspectiveOffCenterFrustum',
-        './FrustumCommands'
+        './FrustumCommands',
+        './SunPostProcess'
     ], function(
         CesiumMath,
         Color,
@@ -62,7 +63,8 @@ define([
         FrameState,
         OrthographicFrustum,
         PerspectiveOffCenterFrustum,
-        FrustumCommands) {
+        FrustumCommands,
+        SunPostProcess) {
     "use strict";
 
     /**
@@ -86,6 +88,8 @@ define([
         this._animations = new AnimationCollection();
 
         this._shaderFrameCount = 0;
+
+        this._sunPostProcess = new SunPostProcess();
 
         this._commandList = [];
         this._frustumCommandsList = [];
@@ -117,6 +121,15 @@ define([
          * @default undefined
          */
         this.skyAtmosphere = undefined;
+
+        /**
+         * The {@link Sun}.
+         *
+         * @type Sun
+         *
+         * @default undefined
+         */
+        this.sun = undefined;
 
         /**
          * The background color, which is only visible if there is no sky box, i.e., {@link Scene#skyBox} is undefined.
@@ -232,10 +245,6 @@ define([
         return this._animations;
     };
 
-    /**
-     * DOC_TBA
-     * @memberof Scene
-     */
     function clearPasses(passes) {
         passes.color = false;
         passes.pick = false;
@@ -315,6 +324,7 @@ define([
 
     var scratchCullingVolume = new CullingVolume();
     var distances = new Interval();
+
     function createPotentiallyVisibleSet(scene, listName) {
         var commandLists = scene._commandList;
         var cullingVolume = scene._frameState.cullingVolume;
@@ -398,17 +408,47 @@ define([
         }
     }
 
+    function isSunVisible(command, frameState) {
+        var occluder = (frameState.mode === SceneMode.SCENE3D) ? frameState.occluder: undefined;
+        var cullingVolume = frameState.cullingVolume;
+
+        // get user culling volume minus the far plane.
+        var planes = scratchCullingVolume.planes;
+        for (var k = 0; k < 5; ++k) {
+            planes[k] = cullingVolume.planes[k];
+        }
+        cullingVolume = scratchCullingVolume;
+
+        return ((typeof command !== 'undefined') &&
+                 ((typeof command.boundingVolume === 'undefined') ||
+                  !command.cull ||
+                  ((cullingVolume.getVisibility(command.boundingVolume) !== Intersect.OUTSIDE) &&
+                   (typeof occluder === 'undefined' || occluder.isBoundingSphereVisible(command.boundingVolume)))));
+    }
+
     function executeCommands(scene, passState) {
+        var frameState = scene._frameState;
         var camera = scene._camera;
         var frustum = camera.frustum.clone();
         var context = scene._context;
         var us = context.getUniformState();
-        var skyBoxCommand = typeof scene.skyBox !== 'undefined' ? scene.skyBox.update(context, scene._frameState) : undefined;
-        var skyAtmosphereCommand = typeof scene.skyAtmosphere !== 'undefined' ? scene.skyAtmosphere.update(context, scene._frameState) : undefined;
+
+        var skyBoxCommand = (frameState.passes.color && typeof scene.skyBox !== 'undefined') ? scene.skyBox.update(context, frameState) : undefined;
+        var skyAtmosphereCommand = (frameState.passes.color && typeof scene.skyAtmosphere !== 'undefined') ? scene.skyAtmosphere.update(context, frameState) : undefined;
+        var sunCommand = (frameState.passes.color && typeof scene.sun !== 'undefined') ? scene.sun.update(context, frameState) : undefined;
+        var sunVisible = isSunVisible(sunCommand, frameState);
+
+        if (sunVisible) {
+            passState.framebuffer = scene._sunPostProcess.update(context);
+        }
 
         var clear = scene._clearColorCommand;
         Color.clone(defaultValue(scene.backgroundColor, Color.BLACK), clear.color);
         clear.execute(context, passState);
+
+        if (sunVisible) {
+            scene._sunPostProcess.clear(context, scene.backgroundColor);
+        }
 
         // Ideally, we would render the sky box and atmosphere last for
         // early-z, but we would have to draw it in each frustum
@@ -422,6 +462,10 @@ define([
 
         if (typeof skyAtmosphereCommand !== 'undefined') {
             skyAtmosphereCommand.execute(context, passState);
+        }
+
+        if (typeof sunCommand !== 'undefined' && sunVisible) {
+            sunCommand.execute(context, passState);
         }
 
         var clearDepthStencil = scene._clearDepthStencilCommand;
@@ -443,6 +487,11 @@ define([
             for (var j = 0; j < length; ++j) {
                 commands[j].execute(context, passState);
             }
+        }
+
+        if (sunVisible) {
+            scene._sunPostProcess.execute(context);
+            passState.framebuffer = undefined;
         }
     }
 
@@ -494,13 +543,14 @@ define([
 
         us.update(frameState);
 
-        this._commandList.length = 0;
-        this._primitives.update(this._context, frameState, this._commandList);
+        var context = this._context;
 
-        var passState = this._passState;
-        passState.framebuffer = undefined;
+        this._commandList.length = 0;
+        this._primitives.update(context, frameState, this._commandList);
 
         createPotentiallyVisibleSet(this, 'colorList');
+
+        var passState = this._passState;
         executeCommands(this, passState);
         executeOverlayCommands(this, passState);
     };
@@ -632,6 +682,8 @@ define([
         this._primitives = this._primitives && this._primitives.destroy();
         this.skyBox = this.skyBox && this.skyBox.destroy();
         this.skyAtmosphere = this.skyAtmosphere && this.skyAtmosphere.destroy();
+        this.sun = this.sun && this.sun.destroy();
+        this._sunPostProcess = this._sunPostProcess && this._sunPostProcess.destroy();
         this._context = this._context && this._context.destroy();
         return destroyObject(this);
     };